requirements.txt
celery==4.2.1
Django==2.1.1
django-heroku==0.3.1
django-taggit==0.23.0
django-widget-tweaks==1.4.3
gunicorn==19.9.0
psycopg2==2.7.5
redis==2.10.6
tld==0.9.1
